rename files to work better with latest node versions
