+++ package.json
@@ -19,7 +19,7 @@
     "format": "prettier --plugin-search-dir . --write .",
     "prepack": "npm run prepack --workspaces",
     "pretest": "npm run build --workspace linkifyjs --silent",
-    "test": "mocha --file test/setup.js --recursive",
+    "test": "mocha --file test/setup.mjs --recursive",
     "test:coverage": "nyc --reporter=lcov --reporter=text mocha test/index.js",
     "test:ci": "run-s test:ci1 test:ci2",
     "test:ci1": "karma start test/ci1.conf.js --single-run",
