+++ packages/linkify-jquery/src/linkify-jquery.mjs
