+++ packages/linkify-plugin-hashtag/src/hashtag.mjs
