+++ packages/linkify-plugin-hashtag/src/index.mjs
