+++ packages/linkify-plugin-mention/src/index.mjs
