+++ packages/linkify-plugin-mention/src/mention.mjs
