+++ packages/linkify-plugin-ticket/src/index.mjs
