+++ packages/linkify-plugin-ticket/src/ticket.mjs
