+++ packages/linkify-react/src/linkify-react.mjs
