+++ packages/linkifyjs/src/assign.mjs
