+++ packages/linkifyjs/src/options.mjs
@@ -1,4 +1,4 @@
-import assign from './assign.js';
+import assign from './assign.mjs';
 
 /**
  * An object where each key is a valid DOM Event Name such as `click` or `focus`
