+++ packages/linkifyjs/src/text.mjs
