+++ test/spec/linkifyjs/parser.test.mjs
@@ -1,7 +1,7 @@
 import { expect } from 'chai';
-import { Text, Url, Email, Nl } from 'linkifyjs/src/multi.js';
-import * as scanner from 'linkifyjs/src/scanner.js';
-import * as parser from 'linkifyjs/src/parser.js';
+import { Text, Url, Email, Nl } from 'linkifyjs/src/multi.mjs';
+import * as scanner from 'linkifyjs/src/scanner.mjs';
+import * as parser from 'linkifyjs/src/parser.mjs';
 
 /**
 	[0] - Original text to parse (should tokenize first)
